Split CacheType from Deltas
